requirements.txt
boto3==1.16.30
botocore==1.19.30
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
confusable-homoglyphs==3.2.0
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.44.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-compressor==2.4
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-debug-toolbar==2.2
django-editorjs==0.2.1
django-environ==0.4.5
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-registration==3.1.1
django-rest-auth==0.9.5
django-storages==1.10.1
django-widgets-improved==1.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
oauthlib==3.1.0
pilkit==2.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
rcssmin==1.0.6
requests==2.25.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
